Componetize
